pyproject.toml
[project]
name = "ai-langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.12",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.2",
]
